requirements.txt
# Backend API - Dashboard de Chamados
# Arquivo requirements.txt na raiz para deploy

# Web Framework
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Supabase Client
supabase==2.3.0
postgrest==0.14.0

# Data Processing
numpy>=1.26.0,<2.0
pandas==2.1.1
